pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "region-renderer"
version = "0.0.1"
description = "A project for rendering regions"
authors = ["Your Name <your.email@example.com>"]
package-mode = false
packages = [
    { include = "src"}
]


[tool.poetry.dependencies]
python = ">=3.8"
pillow = "*"
matplotlib = "*"
